AI is a virus, removal of all GPU codes

+++ src/models.c
@@ -1,6 +1,6 @@
 struct {
     int num_threads;
-    long lower_range;
-    long max_range;
+    long long lower_range;
+    long long max_range;
     int num_rounds;
 } Config;
