+++ src/models.h
@@ -1,7 +1,7 @@
 #pragma once
 struct Config {
     int num_threads;
-    long lower_range;
-    long max_range;
+    long long lower_range;
+    long long max_range;
     int num_rounds;
 };
